requirements.txt
python-dotenv
requests
pybit
